feat: Add Product and Store Services classes
This commit adds the ProductServices and StoreServices classes to the project. These classes provide the necessary services for managing products and stores. The ProductServices class includes methods for adding a product to a store and retrieving products by store ID. The StoreServices class includes methods for adding a store, retrieving all stores, and retrieving a store by ID.

Co-authored-by: daniel-lujan <daniel.lujan@udea.edu.co>
Co-authored-by: GuillermoSanchez11 <guillermo.sanchez1@udea.edu.co>
Co-authored-by: anavegaa <ana.vega@udea.edu.co>
Co-authored-by: MaritzaTC <maritza.tabarezc@udea.edu.co>
Co-authored-by: JuanPablo-A <jpablo.arango1@udea.edu.co>