feat: Add Product and Store Controllers
This commit adds the ProductController and StoreController classes to the project. These classes handle the HTTP requests and responses for managing products and stores. The ProductController includes methods for adding a product to a store and retrieving products by store ID. The StoreController includes methods for adding a store and retrieving all stores.

Co-authored-by: daniel-lujan <daniel.lujan@udea.edu.co>
Co-authored-by: GuillermoSanchez11 <guillermo.sanchez1@udea.edu.co>
Co-authored-by: anavegaa <ana.vega@udea.edu.co>
Co-authored-by: MaritzaTC <maritza.tabarezc@udea.edu.co>
Co-authored-by: JuanPablo-A <jpablo.arango1@udea.edu.co>